+++ src/layout/DefaultLayout/DefaultLayout.test.tsx
@@ -15,7 +15,7 @@ describe('DefaultLayout component options', () => {
       </div>
     )
     renderWithAuthAndApollo(
-      <DefaultLayout displayFeedbackCard={false} rightSidebar={sidebar}>
+      <DefaultLayout rightSidebar={sidebar}>
         <h1>Test Page</h1>
       </DefaultLayout>,
       {},
@@ -70,7 +70,7 @@ describe('DefaultLayout component', () => {
 
   it('renders a feedback link', () => {
     expect(
-      screen.getByRole('link', { name: 'Send us feedback' })
+      screen.getByRole('link', { name: 'feedback@ussforbit.us' })
     ).toHaveAttribute(
       'href',
       'mailto:feedback@ussforbit.us?subject=USSF portal feedback'
@@ -79,7 +79,7 @@ describe('DefaultLayout component', () => {
 
   it('renders common layout elements', () => {
     expect(screen.getAllByRole('banner')).toHaveLength(2) // Gov banner & site header
-    expect(screen.getAllByRole('navigation')).toHaveLength(4) // header, page nav, footer
+    expect(screen.getAllByRole('navigation')).toHaveLength(5) // header, page nav, footer x2, breadcrumbs
   })
 })